Renamed all semantic_framework to semantiva. Created toml file and updated ci pipeline using pdm package
