+++ .github/workflows/ci-pipeline.yaml
@@ -1,13 +1,18 @@
 name: CI Pipeline
 
 on: push
+#    branches:
+#      - main
+#  pull_request:
+#    branches:
+#      - main
 
 env:
-  project_name: semantic_framework # Define a global variable for the project name
-  python_version: '3.10.12' 
+  project_name: semantiva # Define a global variable for the project name
+  python_version: '3.10.12'
 
 jobs:
-  build:
+  linting:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
@@ -16,24 +21,43 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v2
         with:
-          python-version: ${{ env.python_version }} 
+          python-version: ${{ env.python_version }}
 
-      - name: Install dependencies
+      - name: Install PDM
         run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt  # This installs the dependencies from requirements.txt file
+          pip install pdm
+
+      - name: Install dependencies using PDM
+        run: |
+          pdm install
 
       - name: Run pylint
         run: |
-          pylint ${{ env.project_name }} --exit-zero
-
-      # - name: Run mypy
-      #  run: |
-      #    mypy ${{ env.project_name }}
+          pdm run pylint ${{ env.project_name }} --fail-under=7.5
 
       - name: Run black
         run: |
-          black --check ${{ env.project_name }}
+          pdm run black --check ${{ env.project_name }}
+
+  test:
+    runs-on: ubuntu-latest
+    needs: linting  # Ensure that the test job runs after linting job passes
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ env.python_version }}
+
+      - name: Install PDM
+        run: |
+          pip install pdm
+
+      - name: Install dependencies using PDM
+        run: |
+          pdm install
 
       - name: Run pytest
         run: |
