--- requirements.txt
@@ -1,21 +1,0 @@
-astroid==3.3.8
-black==24.10.0
-click==8.1.8
-dill==0.3.9
-exceptiongroup==1.2.2
-iniconfig==2.0.0
-isort==5.13.2
-mccabe==0.7.0
-mypy==1.14.1
-mypy-extensions==1.0.0
-numpy==2.2.1
-packaging==24.2
-pathspec==0.12.1
-pillow==11.1.0
-platformdirs==4.3.6
-pluggy==1.5.0
-pylint==3.3.3
-pytest==8.3.4
-tomli==2.2.1
-tomlkit==0.13.2
-typing_extensions==4.12.2
