+++ semantiva/specializations/image/image_algorithms.py
