+++ semantiva/specializations/image/image_operations.py
@@ -1,4 +1,4 @@
-from semantic_framework.data_operations import DataAlgorithm
+from semantiva.data_operations import DataAlgorithm
 from .image_data_types import ImageDataType, ImageStackDataType
