+++ tests/test_audio_pipeline.py
@@ -1,12 +1,12 @@
 import pytest
 import numpy as np
-from semantic_framework.specializations.audio.audio_data_types import (
+from semantiva.specializations.audio.audio_data_types import (
     SingleChannelAudioDataType,
 )
-from semantic_framework.context_operations import ContextType
-from semantic_framework.payload_operations import Pipeline
+from semantiva.context_operations import ContextType
+from semantiva.payload_operations import Pipeline
 
-from test_audio_algorithm import (
+from .test_audio_algorithm import (
     SingleChannelAudioMultiplyAlgorithm,
     DualChannelAudioMultiplyAlgorithm,
     SingleChannelMockDataProbe,
